+++ src/rule.c
@@ -9,6 +9,8 @@ void rule_serialize(FILE *rsp, struct rule *rule, int index)
             "\t\"label\":\"%s\",\n"
             "\t\"app\":\"%s\",\n"
             "\t\"title\":\"%s\",\n"
+            "\t\"role\":\"%s\",\n"
+            "\t\"subrole\":\"%s\",\n"
             "\t\"display\":%d,\n"
             "\t\"space\":%d,\n"
             "\t\"follow_space\":%s,\n"
@@ -25,6 +27,8 @@ void rule_serialize(FILE *rsp, struct rule *rule, int index)
             rule->label ? rule->label : "",
             rule->app ? rule->app : "",
             rule->title ? rule->title : "",
+            rule->role ? rule->role : "",
+            rule->subrole ? rule->subrole : "",
             display_arrangement(rule->did),
             space_manager_mission_control_index(rule->sid),
             json_bool(rule->follow_space),
@@ -86,9 +90,14 @@ void rule_add(struct rule *rule)
 
 void rule_destroy(struct rule *rule)
 {
-    if (rule->app_regex_valid)   regfree(&rule->app_regex);
-    if (rule->title_regex_valid) regfree(&rule->title_regex);
-    if (rule->label) free(rule->label);
-    if (rule->app)   free(rule->app);
-    if (rule->title) free(rule->title);
+    if (rule->app_regex_valid)     regfree(&rule->app_regex);
+    if (rule->title_regex_valid)   regfree(&rule->title_regex);
+    if (rule->role_regex_valid)    regfree(&rule->role_regex);
+    if (rule->subrole_regex_valid) regfree(&rule->subrole_regex);
+
+    if (rule->label)   free(rule->label);
+    if (rule->app)     free(rule->app);
+    if (rule->title)   free(rule->title);
+    if (rule->role)    free(rule->role);
+    if (rule->subrole) free(rule->subrole);
 }